requirements-ml.txt
joblib==1.5.2
numpy==2.3.4
pandas==2.3.3
scikit-learn==1.7.2
scipy==1.16.3
firebase_admin==7.1.0
google-cloud-firestore==2.21.0
python-dotenv==1.1.1
pydantic==2.12.3